% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_impactsuite_docs.R
\docType{package}
\name{impactsuite-package}
\alias{impactsuite}
\alias{impactsuite-package}
\title{IMPACT suite: meta package installing all packages developed by IMPACT initiatives}
\description{
The suite contains the following packages:
}
\details{
\itemize{
\item stable and in use:
\itemize{
\item msni19
\item Setviz
\item surveyweights
\item questionnaireHTML
\item koboquest
\item xlsformfill
\item cleaninginspectoR
\item kobostandards
\item composr
\item mergekobodata
\item visualisationIMPACT
}
\item experimental:
\itemize{
\item odktidy
\item clog
}
\item HQ use only:
\itemize{
\item researchcyclematrix
}
}
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
